requirements.txt
streamlit
langchain
langchain-community
ollama
lanchain_core
python-dotenv
langserve
fastapi
uvicorn